# exports/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/audit/ops/logs": {
      "get": {
        "description": "按时间范围/操作者/动作类型等条件查询审计日志。",
        "operationId": "AuditController_list",
        "parameters": [
          {
            "name": "timeRange",
            "required": false,
            "in": "query",
            "description": "时间范围，如 24h",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor",
            "required": false,
            "in": "query",
            "description": "操作者",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actionType",
            "required": false,
            "in": "query",
            "description": "动作类型",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target",
            "required": false,
            "in": "query",
            "description": "目标名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "条数，默认50",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "审计列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpsAuditLogResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询操作审计日志",
        "tags": [
          "Audit"
        ]
      }
    },
    "/audit/ops/logs/{id}": {
      "get": {
        "description": "返回单条审计日志详情（含 steps/rollbackSteps 摘要）。",
        "operationId": "AuditController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "审计日志 ID",
            "schema": {
              "example": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "审计详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsAuditLogResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "审计详情",
        "tags": [
          "Audit"
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "description": "返回 access/refresh 双 token 与用户信息，用于后续受保护接口的 Bearer 鉴权。",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "token 信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "登录获取 access/refresh token（内置 demo 用户：admin/admin、user/user；可通过 env 覆盖）",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "refreshToken 有效时颁发新 token；用于无感续期。",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "新的 token 信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "使用 refreshToken 换取新的 access/refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "返回当前登录用户的基础信息与角色。",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取当前用户信息（需 Bearer Token）",
        "tags": [
          "Auth"
        ]
      }
    },
    "/alerts/config": {
      "get": {
        "description": "读取当前告警开关、收件人列表与阈值配置。",
        "operationId": "AlertsController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "告警配置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取告警配置",
        "tags": [
          "Alerts"
        ]
      },
      "put": {
        "description": "更新告警开关、收件人列表与阈值配置，并记录审计。",
        "operationId": "AlertsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的告警配置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新告警配置",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/alerts/send": {
      "post": {
        "description": "根据当前配置发送告警邮件，写入发送记录与审计。",
        "operationId": "AlertsController_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendAlertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendAlertResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "手动触发告警（演练/测试）",
        "tags": [
          "Alerts"
        ]
      }
    }
  },
  "info": {
    "title": "WAF 控制面 API",
    "description": "v1 接口文档",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OpsAuditLogResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "日志 ID",
            "example": 100
          },
          "actionType": {
            "type": "string",
            "description": "动作类型",
            "example": "PUBLISH_POLICY"
          },
          "targetType": {
            "type": "string",
            "description": "目标类型",
            "example": "server"
          },
          "targetName": {
            "type": "string",
            "description": "目标名称",
            "example": "example.com"
          },
          "status": {
            "type": "string",
            "description": "执行状态",
            "enum": [
              "SUCCESS",
              "FAILED"
            ]
          },
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "admin"
          },
          "note": {
            "type": "string",
            "description": "备注"
          },
          "detail": {
            "type": "object",
            "description": "细节（steps/rollbackSteps 等摘要）",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间（ISO）"
          }
        },
        "required": [
          "id",
          "actionType",
          "targetType",
          "status",
          "createdAt"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户 ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "admin"
          },
          "role": {
            "type": "string",
            "description": "角色",
            "enum": [
              "admin",
              "user"
            ]
          }
        },
        "required": [
          "id",
          "username",
          "role"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Access Token（JWT）"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh Token（JWT）"
          },
          "expiresIn": {
            "type": "string",
            "description": "Access Token 过期时间（如 15m）"
          },
          "refreshExpiresIn": {
            "type": "string",
            "description": "Refresh Token 过期时间（如 7d）"
          },
          "user": {
            "description": "用户信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthUserDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "refreshExpiresIn",
          "user"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AlertThresholdsDto": {
        "type": "object",
        "properties": {
          "blockRate": {
            "type": "number",
            "description": "拦截率阈值，0-1 之间小数",
            "example": 0.5
          },
          "qps": {
            "type": "number",
            "description": "QPS 阈值",
            "example": 200
          },
          "attackTypeCounts": {
            "type": "object",
            "description": "按攻击类型的计数阈值，key 为攻击类型",
            "example": {
              "SQL_INJECTION": 10,
              "XSS": 5
            }
          }
        }
      },
      "AlertConfigResponse": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "是否开启告警",
            "example": true
          },
          "emails": {
            "description": "收件人邮箱列表",
            "example": [
              "ops@example.com"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thresholds": {
            "description": "阈值配置",
            "allOf": [
              {
                "$ref": "#/components/schemas/AlertThresholdsDto"
              }
            ]
          },
          "updatedBy": {
            "type": "string",
            "description": "最近更新人",
            "example": "admin"
          },
          "updatedAt": {
            "type": "string",
            "description": "最近更新时间",
            "example": "2026-01-03T01:23:45Z"
          }
        },
        "required": [
          "enabled",
          "emails",
          "thresholds"
        ]
      },
      "UpdateAlertConfigDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "告警开关",
            "default": true
          },
          "emails": {
            "description": "收件人邮箱列表",
            "example": [
              "ops@example.com"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thresholds": {
            "type": "object",
            "description": "触发阈值（拦截率/QPS/攻击类型计数）",
            "example": {
              "blockRate": 0.5,
              "qps": 200,
              "attackTypeCounts": {
                "SQL_INJECTION": 10
              }
            }
          }
        }
      },
      "SendAlertDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "告警标题",
            "example": "演练告警：高拦截率"
          },
          "content": {
            "type": "string",
            "description": "告警内容",
            "example": "近5分钟拦截率 > 50%"
          }
        },
        "required": [
          "subject"
        ]
      },
      "SendAlertResponse": {
        "type": "object",
        "properties": {
          "sent": {
            "type": "boolean",
            "description": "是否已发送（需开启且有收件人）"
          },
          "subject": {
            "type": "string",
            "description": "标题"
          },
          "content": {
            "type": "string",
            "description": "正文"
          },
          "recipients": {
            "description": "收件人列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sentAt": {
            "type": "string",
            "description": "发送时间 ISO 字符串"
          },
          "error": {
            "type": "string",
            "description": "失败原因（未开启、无收件人或 SMTP 错误）"
          }
        },
        "required": [
          "sent",
          "subject",
          "content",
          "recipients",
          "sentAt"
        ]
      }
    }
  }
}